# Enable crypto HW acceleration for STM32F437xG platforms