Enable crypto HW acceleration for STM43F437xG platforms, i.e. ublox C030 family boards.

`mbed-os-example-tls/benchmark/` before:

```
  SHA-256                  :       2936 KB/s
  SHA-512                  :        886 KB/s
  AES-CBC-128              :       2315 KB/s
  AES-CBC-192              :       2031 KB/s
  AES-CBC-256              :       1810 KB/s
  AES-GCM-128              :        708 KB/s
  AES-GCM-192              :        679 KB/s
  AES-GCM-256              :        652 KB/s
  AES-CCM-128              :        957 KB/s
  AES-CCM-192              :        857 KB/s
  AES-CCM-256              :        776 KB/s
  CTR_DRBG (NOPR)          :       1922 KB/s
  CTR_DRBG (PR)            :       1361 KB/s
  HMAC_DRBG SHA-256 (NOPR) :        337 KB/s
  HMAC_DRBG SHA-256 (PR)   :        296 KB/s
  RSA-2048                 :      20 ms/ public
  RSA-2048                 :     782 ms/private
  RSA-4096                 :      68 ms/ public
  RSA-4096                 :    4160 ms/private
  ECDHE-secp384r1          :     696 ms/handshake
  ECDHE-secp256r1          :     461 ms/handshake
  ECDHE-Curve25519         :     386 ms/handshake
  ECDH-secp384r1           :     343 ms/handshake
  ECDH-secp256r1           :     232 ms/handshake
  ECDH-Curve25519          :     192 ms/handshake
```
....and after:

```
  SHA-256                  :      15081 KB/s
  SHA-512                  :        886 KB/s
  AES-CBC-128              :      31542 KB/s
  AES-CBC-192              :      31542 KB/s
  AES-CBC-256              :      31542 KB/s
  AES-GCM-128              :        913 KB/s
  AES-GCM-192              :        913 KB/s
  AES-GCM-256              :        903 KB/s
  AES-CCM-128              :       2172 KB/s
  AES-CCM-192              :       2172 KB/s
  AES-CCM-256              :       2170 KB/s
  CTR_DRBG (NOPR)          :       8512 KB/s
  CTR_DRBG (PR)            :       5244 KB/s
  HMAC_DRBG SHA-256 (NOPR) :        470 KB/s
  HMAC_DRBG SHA-256 (PR)   :        411 KB/s
  RSA-2048                 :      20 ms/ public
  RSA-2048                 :     781 ms/private
  RSA-4096                 :      68 ms/ public
  RSA-4096                 :    4054 ms/private
  ECDHE-secp384r1          :     692 ms/handshake
  ECDHE-secp256r1          :     461 ms/handshake
  ECDHE-Curve25519         :     385 ms/handshake
  ECDH-secp384r1           :     339 ms/handshake
  ECDH-secp256r1           :     232 ms/handshake
  ECDH-Curve25519          :     199 ms/handshake
```

Test results for `mbed-os-tests-mbedtls-selftest`:

```
+-------------------------+-----------------+--------------------------------+---------------------------+--------+--------+--------+--------------------+
| target                  | platform_name   | test suite                     | test case                 | passed | failed | result | elapsed_time (sec) |
+-------------------------+-----------------+--------------------------------+---------------------------+--------+--------+--------+--------------------+
| UBLOX_C030_U201-GCC_ARM | UBLOX_C030_U201 | mbed-os-tests-mbedtls-selftest | mbedtls_entropy_self_test | 1      | 0      | OK     | 0.06               |
| UBLOX_C030_U201-GCC_ARM | UBLOX_C030_U201 | mbed-os-tests-mbedtls-selftest | mbedtls_sha256_self_test  | 1      | 0      | OK     | 0.11               |
| UBLOX_C030_U201-GCC_ARM | UBLOX_C030_U201 | mbed-os-tests-mbedtls-selftest | mbedtls_sha512_self_test  | 1      | 0      | OK     | 1.98               |
+-------------------------+-----------------+--------------------------------+---------------------------+--------+--------+--------+--------------------+
```
NOTE: this should only be merged once PR #5018 (which fixes an AES HW crypto driver issue) has been merged.