@0xc0170 2017-08-15
@RonEld @andresag01 please review tihs addition