@andresag01 2017-08-18
@yanesca, @0xc0170: There was a recent bug report for AES hardware acceleration on STM32F439xI: https://github.com/ARMmbed/mbed-os/issues/4928. This PR looks good to me, but given that STM32F439xI and STM32F437xG share the same `aes_alt.h` and `aes_alt.c` in [this folder](https://github.com/ARMmbed/mbed-os/tree/master/features/mbedtls/targets/TARGET_STM) I suggest we test the change using the tls-client before merging.