@0xc0170 2017-08-21
> @yanesca, @0xc0170: There was a recent bug report for AES hardware acceleration on STM32F439xI: #4928. This PR looks good to me, but given that STM32F439xI and STM32F437xG share the same aes_alt.h and aes_alt.c in this folder I suggest we test the change using the tls-client before merging.

@RobMeades Can you confirm ?