@RobMeades 2017-08-21
@0xc0170 I only ran the benchmarks from within [mbed-os-example-tls](https://github.com/ARMmbed/mbed-os-example-tls) and these seemed to run as expected.  I can confirm, though, that the HW crypto source files are shared between all STM32F4 devices, so if there's an issue it should be common.