@0xc0170 2017-09-01
> @yanesca, @0xc0170: There was a recent bug report for AES hardware acceleration on STM32F439xI: #4928. This PR looks good to me, but given that STM32F439xI and STM32F437xG share the same aes_alt.h and aes_alt.c in this folder I suggest we test the change using the tls-client before merging.

@RobMeades Is that correct? You above test says otherwise. We are currently disabling AES for those platform until it's fixed. If they share the implementation, how come this one does not fail?

@andresag01 I noticed you are happy with testing above, thus approved this addition?