@RobMeades 2017-09-01
The HW crypto source files are certainly shared between all STM32F4 devices and I _was_ able to pass the test.  Maybe there is some other factor involved in causing the failure?