@0xc0170 2017-09-01
> The HW crypto source files are certainly shared between all STM32F4 devices and I was able to pass the test. Maybe there is some other factor involved in causing the failure?
 
that is what is unknown to me. To be certain, I would like to have one more test run on this target. @andresag01 @yanesca @RonEld What is your stand? Can you run quickly the test above on this platform to confirm the passing the test to accept this addition