@yanesca 2017-09-01
I am sorry, but I still need to review this before we can accept it and I am even more sorry that I couldn't review this yet. Unfortunately at this point I have no estimate when can I do the review, but I will do it as soon as I can!

@RobMeades When you say that this has passed the test, do you mean compiling and successfully running on target the `tls-client` application from the https://github.com/ARMmbed/mbed-os-example-tls repository?