@RobMeades 2017-09-01
@yanesca: yes, that's correct, with the change in #4934 removed so I am using HW crypto.  I'll try it again from scratch on Monday just to be sure.