@RobMeades 2017-09-04
Well, it seems I was wrong.  I've retrieved the latest [https://github.com/ARMmbed/mbed-os-example-tls](https://github.com/ARMmbed/mbed-os-example-tls), made sure that `tls-client` has the latest `mbed-os`, added this change and then compiled and run the `tls-client` example on my STM32F437 target.  If I have HW crypto on for AES it reports:
```
Using Ethernet LWIP
Client IP Address is 10.38.20.210
Connecting with developer.mbed.org
Starting the TLS handshake...
mbedtls_ssl_handshake() failed: -0x7780 (-30592): SSL - A fatal alert message was received from our peer
```

...and stops, while if I switch off HW crypto for AES it reports:

```
Using Ethernet LWIP
Client IP Address is 10.38.20.210
Connecting with developer.mbed.org
Starting the TLS handshake...
TLS connection to developer.mbed.org established
.etc..```

...and ends with "Hello world!"

So I guess this is the issue reported in #4928.  Over in #4928 I see that Martin comments: "The fix will get CI once CI is back running", but I can't see anything in the comments to indicate that there is a fix.  If there is one in the pipeline, this PR can wait for it.  If there is not, I will re-submit this PR with `MBEDTLS_AES_ALT` left out.