@RobMeades 2017-09-05
I suggest that we only merge this change once #5018 has been merged, I will update the description to clarify that.