@ciarmcom 2017-09-07
ARM Internal Ref: IOTSSL-1720