@0xc0170 2017-09-28
> I suggest that we only merge this change once #5018 has been merged, I will update the description to clarify that.

That PR was merged. What is remaining here?