@andresag01 2017-09-28
@0xc0170: We should wait for @yanesca to approve the changes as well before merging.